add class to tidy
